gold_chatbot/requirements.txt
faiss-cpu
pandas
numpy
langchain
langchain-google-genai
langchain-community
google-generativeai
pypdf
langgraph
python-dotenv
psycopg2-binary
fastapi
uvicorn
pydantic
langchain-ollama

#frontend
streamlit
pandas
openai
requests
